third_party/blink/web_tests/external/wpt/css/css-grid/masonry/tentative/items/row-minimum-size-grid-items-002-expected.txt

This is a testharness.js-based test.
Found 60 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] .masonry 1
  assert_equals: \n<div class="masonry" style="grid-template-rows: auto;">\n  <div class="verticalLR" data-expected-height="100">XXXXXXXXXX</div>\n  <div data-expected-height="100"></div>\n</div>\nheight expected 100 but got 0
[FAIL] .masonry 2
  assert_equals: \n<div class="masonry" style="grid-template-rows: 0px;">\n  <div class="verticalLR" data-expected-height="0">XXXXXXXXXX</div>\n  <div data-expected-height="0"></div>\n</div>\nheight expected 0 but got 100
[FAIL] .masonry 3
  assert_equals: \n<div class="masonry" style="grid-template-rows: 25px;">\n  <div class="verticalLR" data-expected-height="25">XXXXXXXXXX</div>\n  <div data-expected-height="25"></div>\n</div>\nheight expected 25 but got 100
[FAIL] .masonry 4
  assert_equals: \n<div class="masonry" style="grid-template-rows: minmax(auto, 0px);">\n  <div class="verticalLR" data-expected-height="0">XXXXXXXXXX</div>\n  <div data-expected-height="0"></div>\n</div>\nheight expected 0 but got 100
[FAIL] .masonry 5
  assert_equals: \n<div class="masonry" style="grid-template-rows: minmax(auto, 25px);">\n  <div class="verticalLR" data-expected-height="25">XXXXXXXXXX</div>\n  <div data-expected-height="25"></div>\n</div>\nheight expected 25 but got 100
[FAIL] .masonry 6
  assert_equals: \n<div class="masonry" style="grid-template-rows: minmax(auto, 0px);">\n  <div class="verticalLR" data-expected-height="10" style="height: 10px;">XXXXXXXXXX</div>\n  <div data-expected-height="10"></div>\n</div>\nheight expected 10 but got 0
[FAIL] .masonry 7
  assert_equals: \n<div class="masonry" style="grid-template-rows: minmax(auto, 25px);">\n  <div class="verticalLR" data-expected-height="10" style="height: 10px;">XXXXXXXXXX</div>\n  <div data-expected-height="25"></div>\n</div>\nheight expected 25 but got 0
[FAIL] .masonry 8
  assert_equals: \n<div class="masonry" style="grid-template-rows: minmax(auto, 0px);">\n  <div class="verticalLR" data-expected-height="0" style="margin: 5px 0px;"></div>\n  <div data-expected-height="10"></div>\n</div>\nheight expected 10 but got 0
[FAIL] .masonry 9
  assert_equals: \n<div class="masonry" style="grid-template-rows: minmax(auto, 25px);">\n  <div class="verticalLR" data-expected-height="15" style="margin: 5px 0px;"></div>\n  <div data-expected-height="25"></div>\n</div>\nheight expected 15 but got 0
[FAIL] .masonry 10
  assert_equals: \n<div class="masonry" style="grid-template-rows: minmax(auto, 0px);">\n  <div class="verticalLR" data-expected-height="10" style="padding: 5px 0px;"></div>\n  <div data-expected-height="10"></div>\n</div>\nheight expected 10 but got 0
[FAIL] .masonry 11
  assert_equals: \n<div class="masonry" style="grid-template-rows: minmax(auto, 25px);">\n  <div class="verticalLR" data-expected-height="25" style="padding: 5px 0px;"></div>\n  <div data-expected-height="25"></div>\n</div>\nheight expected 25 but got 10
[FAIL] .masonry 12
  assert_equals: \n<div class="masonry" style="grid-template-rows: minmax(auto, 0px);">\n  <div class="verticalLR" data-expected-height="10" style="border: solid 5px blue;"></div>\n  <div data-expected-height="10"></div>\n</div>\nheight expected 10 but got 0
[FAIL] .masonry 13
  assert_equals: \n<div class="masonry" style="grid-template-rows: minmax(auto, 25px);">\n  <div class="verticalLR" data-expected-height="25" style="border: solid 5px blue;"></div>\n  <div data-expected-height="25"></div>\n</div>\nheight expected 25 but got 10
[FAIL] .masonry 14
  assert_equals: \n<div class="masonry" style="grid-template-rows: minmax(auto, 0px);">\n  <div class="verticalLR" data-expected-height="8" style="height: 4px; margin: 1px; padding: 1px; border: solid 1px blue;"></div>\n  <div data-expected-height="10"></div>\n</div>\nheight expected 10 but got 0
[FAIL] .masonry 15
  assert_equals: \n<div class="masonry" style="grid-template-rows: minmax(auto, 25px);">\n  <div class="verticalLR" data-expected-height="8" style="height: 4px; margin: 1px; padding: 1px; border: solid 1px blue;"></div>\n  <div data-expected-height="25"></div>\n</div>\nheight expected 25 but got 0
[FAIL] .masonry 16
  assert_equals: \n<div class="masonry" style="grid-template-rows: auto;">\n  <div class="verticalRL" data-expected-height="100">XXXXXXXXXX</div>\n  <div data-expected-height="100"></div>\n</div>\nheight expected 100 but got 0
[FAIL] .masonry 17
  assert_equals: \n<div class="masonry" style="grid-template-rows: 0px;">\n  <div class="verticalRL" data-expected-height="0">XXXXXXXXXX</div>\n  <div data-expected-height="0"></div>\n</div>\nheight expected 0 but got 100
[FAIL] .masonry 18
  assert_equals: \n<div class="masonry" style="grid-template-rows: 25px;">\n  <div class="verticalRL" data-expected-height="25">XXXXXXXXXX</div>\n  <div data-expected-height="25"></div>\n</div>\nheight expected 25 but got 100
[FAIL] .masonry 19
  assert_equals: \n<div class="masonry" style="grid-template-rows: minmax(auto, 0px);">\n  <div class="verticalRL" data-expected-height="0">XXXXXXXXXX</div>\n  <div data-expected-height="0"></div>\n</div>\nheight expected 0 but got 100
[FAIL] .masonry 20
  assert_equals: \n<div class="masonry" style="grid-template-rows: minmax(auto, 25px);">\n  <div class="verticalRL" data-expected-height="25">XXXXXXXXXX</div>\n  <div data-expected-height="25"></div>\n</div>\nheight expected 25 but got 100
[FAIL] .masonry 21
  assert_equals: \n<div class="masonry" style="grid-template-rows: minmax(auto, 0px);">\n  <div class="verticalRL" data-expected-height="10" style="height: 10px;">XXXXXXXXXX</div>\n  <div data-expected-height="10"></div>\n</div>\nheight expected 10 but got 0
[FAIL] .masonry 22
  assert_equals: \n<div class="masonry" style="grid-template-rows: minmax(auto, 25px);">\n  <div class="verticalRL" data-expected-height="10" style="height: 10px;">XXXXXXXXXX</div>\n  <div data-expected-height="25"></div>\n</div>\nheight expected 25 but got 0
[FAIL] .masonry 23
  assert_equals: \n<div class="masonry" style="grid-template-rows: minmax(auto, 0px);">\n  <div class="verticalRL" data-expected-height="0" style="margin: 5px 0px;"></div>\n  <div data-expected-height="10"></div>\n</div>\nheight expected 10 but got 0
[FAIL] .masonry 24
  assert_equals: \n<div class="masonry" style="grid-template-rows: minmax(auto, 25px);">\n  <div class="verticalRL" data-expected-height="15" style="margin: 5px 0px;"></div>\n  <div data-expected-height="25"></div>\n</div>\nheight expected 15 but got 0
[FAIL] .masonry 25
  assert_equals: \n<div class="masonry" style="grid-template-rows: minmax(auto, 0px);">\n  <div class="verticalRL" data-expected-height="10" style="padding: 5px 0px;"></div>\n  <div data-expected-height="10"></div>\n</div>\nheight expected 10 but got 0
[FAIL] .masonry 26
  assert_equals: \n<div class="masonry" style="grid-template-rows: minmax(auto, 25px);">\n  <div class="verticalRL" data-expected-height="25" style="padding: 5px 0px;"></div>\n  <div data-expected-height="25"></div>\n</div>\nheight expected 25 but got 10
[FAIL] .masonry 27
  assert_equals: \n<div class="masonry" style="grid-template-rows: minmax(auto, 0px);">\n  <div class="verticalRL" data-expected-height="10" style="border: solid 5px blue;"></div>\n  <div data-expected-height="10"></div>\n</div>\nheight expected 10 but got 0
[FAIL] .masonry 28
  assert_equals: \n<div class="masonry" style="grid-template-rows: minmax(auto, 25px);">\n  <div class="verticalRL" data-expected-height="25" style="border: solid 5px blue;"></div>\n  <div data-expected-height="25"></div>\n</div>\nheight expected 25 but got 10
[FAIL] .masonry 29
  assert_equals: \n<div class="masonry" style="grid-template-rows: minmax(auto, 0px);">\n  <div class="verticalRL" data-expected-height="8" style="height: 4px; margin: 1px; padding: 1px; border: solid 1px blue;"></div>\n  <div data-expected-height="10"></div>\n</div>\nheight expected 10 but got 0
[FAIL] .masonry 30
  assert_equals: \n<div class="masonry" style="grid-template-rows: minmax(auto, 25px);">\n  <div class="verticalRL" data-expected-height="8" style="height: 4px; margin: 1px; padding: 1px; border: solid 1px blue;"></div>\n  <div data-expected-height="25"></div>\n</div>\nheight expected 25 but got 0
[FAIL] .masonry 31
  assert_equals: \n<div class="masonry verticalLR" style="grid-template-rows: auto;">\n  <div class="horizontalTB" data-expected-width="100">XXXXXXXXXX</div>\n  <div data-expected-width="100"></div>\n</div>\nwidth expected 100 but got 0
[FAIL] .masonry 32
  assert_equals: \n<div class="masonry verticalLR" style="grid-template-rows: minmax(auto, 0px);">\n  <div class="horizontalTB" data-expected-width="0">XXXXXXXXXX</div>\n  <div data-expected-width="0"></div>\n</div>\nwidth expected 0 but got 100
[FAIL] .masonry 33
  assert_equals: \n<div class="masonry verticalLR" style="grid-template-rows: 25px;">\n  <div class="horizontalTB" data-expected-width="25">XXXXXXXXXX</div>\n  <div data-expected-width="25"></div>\n</div>\nwidth expected 25 but got 100
[FAIL] .masonry 34
  assert_equals: \n<div class="masonry verticalLR" style="grid-template-rows: minmax(auto, 0px);">\n  <div class="horizontalTB" data-expected-width="0">XXXXXXXXXX</div>\n  <div data-expected-width="0"></div>\n</div>\nwidth expected 0 but got 100
[FAIL] .masonry 35
  assert_equals: \n<div class="masonry verticalLR" style="grid-template-rows: minmax(auto, 25px);">\n  <div class="horizontalTB" data-expected-width="25">XXXXXXXXXX</div>\n  <div data-expected-width="25"></div>\n</div>\nwidth expected 25 but got 100
[FAIL] .masonry 36
  assert_equals: \n<div class="masonry verticalLR" style="grid-template-rows: minmax(auto, 0px);">\n  <div class="horizontalTB" data-expected-width="10" style="width: 10px;">XXXXXXXXXX</div>\n  <div data-expected-width="10"></div>\n</div>\nwidth expected 10 but got 0
[FAIL] .masonry 37
  assert_equals: \n<div class="masonry verticalLR" style="grid-template-rows: minmax(auto, 25px);">\n  <div class="horizontalTB" data-expected-width="10" style="width: 10px;">XXXXXXXXXX</div>\n  <div data-expected-width="25"></div>\n</div>\nwidth expected 25 but got 0
[FAIL] .masonry 38
  assert_equals: \n<div class="masonry verticalLR" style="grid-template-rows: minmax(auto, 0px);">\n  <div class="horizontalTB" data-expected-width="0" style="margin: 0px 5px;"></div>\n  <div data-expected-width="10"></div>\n</div>\nwidth expected 10 but got 0
[FAIL] .masonry 39
  assert_equals: \n<div class="masonry verticalLR" style="grid-template-rows: minmax(auto, 25px);">\n  <div class="horizontalTB" data-expected-width="15" style="margin: 0px 5px;"></div>\n  <div data-expected-width="25"></div>\n</div>\nwidth expected 15 but got 0
[FAIL] .masonry 40
  assert_equals: \n<div class="masonry verticalLR" style="grid-template-rows: minmax(auto, 0px);">\n  <div class="horizontalTB" data-expected-width="10" style="padding: 0px 5px;"></div>\n  <div data-expected-width="10"></div>\n</div>\nwidth expected 10 but got 0
[FAIL] .masonry 41
  assert_equals: \n<div class="masonry verticalLR" style="grid-template-rows: minmax(auto, 25px);">\n  <div class="horizontalTB" data-expected-width="25" style="padding: 0px 5px;"></div>\n  <div data-expected-width="25"></div>\n</div>\nwidth expected 25 but got 10
[FAIL] .masonry 42
  assert_equals: \n<div class="masonry verticalLR" style="grid-template-rows: minmax(auto, 0px);">\n  <div class="horizontalTB" data-expected-width="10" style="border: solid 5px blue;"></div>\n  <div data-expected-width="10"></div>\n</div>\nwidth expected 10 but got 0
[FAIL] .masonry 43
  assert_equals: \n<div class="masonry verticalLR" style="grid-template-rows: minmax(auto, 25px);">\n  <div class="horizontalTB" data-expected-width="25" style="border: solid 5px blue;"></div>\n  <div data-expected-width="25"></div>\n</div>\nwidth expected 25 but got 10
[FAIL] .masonry 44
  assert_equals: \n<div class="masonry verticalLR" style="grid-template-rows: minmax(auto, 0px);">\n  <div class="horizontalTB" data-expected-width="8" style="width: 4px; margin: 1px; padding: 1px; border: solid 1px blue;"></div>\n  <div data-expected-width="10"></div>\n</div>\nwidth expected 10 but got 0
[FAIL] .masonry 45
  assert_equals: \n<div class="masonry verticalLR" style="grid-template-rows: minmax(auto, 25px);">\n  <div class="horizontalTB" data-expected-width="8" style="width: 4px; margin: 1px; padding: 1px; border: solid 1px blue;"></div>\n  <div data-expected-width="25"></div>\n</div>\nwidth expected 25 but got 0
[FAIL] .masonry 46
  assert_equals: \n<div class="masonry verticalRL" style="grid-template-rows: auto;">\n  <div class="horizontalTB" data-expected-width="100">XXXXXXXXXX</div>\n  <div data-expected-width="100"></div>\n</div>\nwidth expected 100 but got 0
[FAIL] .masonry 47
  assert_equals: \n<div class="masonry verticalRL" style="grid-template-rows: minmax(auto, 0px);">\n  <div class="horizontalTB" data-expected-width="0">XXXXXXXXXX</div>\n  <div data-expected-width="0"></div>\n</div>\nwidth expected 0 but got 100
[FAIL] .masonry 48
  assert_equals: \n<div class="masonry verticalRL" style="grid-template-rows: 25px;">\n  <div class="horizontalTB" data-expected-width="25">XXXXXXXXXX</div>\n  <div data-expected-width="25"></div>\n</div>\nwidth expected 25 but got 100
[FAIL] .masonry 49
  assert_equals: \n<div class="masonry verticalRL" style="grid-template-rows: minmax(auto, 0px);">\n  <div class="horizontalTB" data-expected-width="0">XXXXXXXXXX</div>\n  <div data-expected-width="0"></div>\n</div>\nwidth expected 0 but got 100
[FAIL] .masonry 50
  assert_equals: \n<div class="masonry verticalRL" style="grid-template-rows: minmax(auto, 25px);">\n  <div class="horizontalTB" data-expected-width="25">XXXXXXXXXX</div>\n  <div data-expected-width="25"></div>\n</div>\nwidth expected 25 but got 100
[FAIL] .masonry 51
  assert_equals: \n<div class="masonry verticalRL" style="grid-template-rows: minmax(auto, 0px);">\n  <div class="horizontalTB" data-expected-width="10" style="width: 10px;">XXXXXXXXXX</div>\n  <div data-expected-width="10"></div>\n</div>\nwidth expected 10 but got 0
[FAIL] .masonry 52
  assert_equals: \n<div class="masonry verticalRL" style="grid-template-rows: minmax(auto, 25px);">\n  <div class="horizontalTB" data-expected-width="10" style="width: 10px;">XXXXXXXXXX</div>\n  <div data-expected-width="25"></div>\n</div>\nwidth expected 25 but got 0
[FAIL] .masonry 53
  assert_equals: \n<div class="masonry verticalRL" style="grid-template-rows: minmax(auto, 0px);">\n  <div class="horizontalTB" data-expected-width="0" style="margin: 0px 5px;"></div>\n  <div data-expected-width="10"></div>\n</div>\nwidth expected 10 but got 0
[FAIL] .masonry 54
  assert_equals: \n<div class="masonry verticalRL" style="grid-template-rows: minmax(auto, 25px);">\n  <div class="horizontalTB" data-expected-width="15" style="margin: 0px 5px;"></div>\n  <div data-expected-width="25"></div>\n</div>\nwidth expected 15 but got 0
[FAIL] .masonry 55
  assert_equals: \n<div class="masonry verticalRL" style="grid-template-rows: minmax(auto, 0px);">\n  <div class="horizontalTB" data-expected-width="10" style="padding: 0px 5px;"></div>\n  <div data-expected-width="10"></div>\n</div>\nwidth expected 10 but got 0
[FAIL] .masonry 56
  assert_equals: \n<div class="masonry verticalRL" style="grid-template-rows: minmax(auto, 25px);">\n  <div class="horizontalTB" data-expected-width="25" style="padding: 0px 5px;"></div>\n  <div data-expected-width="25"></div>\n</div>\nwidth expected 25 but got 10
[FAIL] .masonry 57
  assert_equals: \n<div class="masonry verticalRL" style="grid-template-rows: minmax(auto, 0px);">\n  <div class="horizontalTB" data-expected-width="10" style="border: solid 5px blue;"></div>\n  <div data-expected-width="10"></div>\n</div>\nwidth expected 10 but got 0
[FAIL] .masonry 58
  assert_equals: \n<div class="masonry verticalRL" style="grid-template-rows: minmax(auto, 25px);">\n  <div class="horizontalTB" data-expected-width="25" style="border: solid 5px blue;"></div>\n  <div data-expected-width="25"></div>\n</div>\nwidth expected 25 but got 10
[FAIL] .masonry 59
  assert_equals: \n<div class="masonry verticalRL" style="grid-template-rows: minmax(auto, 0px);">\n  <div class="horizontalTB" data-expected-width="8" style="width: 4px; margin: 1px; padding: 1px; border: solid 1px blue;"></div>\n  <div data-expected-width="10"></div>\n</div>\nwidth expected 10 but got 0
[FAIL] .masonry 60
  assert_equals: \n<div class="masonry verticalRL" style="grid-template-rows: minmax(auto, 25px);">\n  <div class="horizontalTB" data-expected-width="8" style="width: 4px; margin: 1px; padding: 1px; border: solid 1px blue;"></div>\n  <div data-expected-width="25"></div>\n</div>\nwidth expected 25 but got 0
Harness: the test ran to completion.
